enh: add tests to docstring and integrate to overhaul
Closes: nipreps/nifreeze#332.
